pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deepfake-detection-fusion"
version = "0.1.0"
description = "Deepfake image detection with spatial-frequency fusion, explainability, robustness, and multi-dataset evaluation"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  # core deep learning
  "timm>=1.0.25",

  # data / math
  "numpy>=2.0",
  "scipy>=1.13",
  "pandas>=2.2",
  "scikit-learn>=1.5",

  # image processing / augmentation
  "pillow>=10.4",
  "opencv-python>=4.10",
  "albumentations>=2.0",

  # visualization / experiment logging
  "matplotlib>=3.8",
  "seaborn>=0.13",
  "tqdm>=4.66",
  "rich>=13.7",
  "wandb>=0.25",

  # config / utility
  "PyYAML>=6.0",
  "python-dotenv>=1.0",

  # explainability
  "grad-cam>=1.5.5",
  "captum>=0.7",

  # robustness / adversarial experiments
  "torchattacks>=3.5.1",

  # dataset download / api
  "datasets>=2.19",
  "kagglehub>=0.3",
]

[project.optional-dependencies]
dev = [
  "jupyterlab>=4.2",
  "ipykernel>=6.29",
  "notebook>=7.2",
  "pytest>=8.2",
  "ruff>=0.6",
  "black>=24.8",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]